% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cleveland}
\alias{cleveland}
\title{Example dataset based on cross-validated prediction of outcome in the Cleveland Heart Study}
\format{
A data frame with 297 rows and three variables
\describe{
  \item{prior.p}{prior probabilities of case status}
  \item{posterior.p}{posterior probabilites of case status}
  \item{y}{case-control status}
}
}
\source{
\url{http://www.homepages.ed.ac.uk/pmckeigu/preprints/classify/demoplotw.html}
}
\usage{
cleveland
}
\description{
Example dataset based on cross-validated prediction of outcome in the Cleveland Heart Study
}
\keyword{datasets}
