% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdists.R
\name{means.densities}
\alias{means.densities}
\title{Means of densities in cases and controls}
\usage{
means.densities(densities)
}
\arguments{
\item{densities}{Adjusted densities computed by
\code{\link{Wdensities.fromraw}}.}
}
\value{
numeric vector of length 2: mean densities in controls and in cases.
}
\description{
Means of densities in cases and controls
}
\seealso{
\code{\link{lambda.model}}
}
