% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotcumfreqs}
\alias{plotcumfreqs}
\title{Plot the cumulative frequency distributions in cases and in controls}
\usage{
plotcumfreqs(densities)
}
\arguments{
\item{densities}{Densities object produced by \code{\link{Wdensities}}.}
}
\value{
A ggplot object representing the cumulative frequency distributions of the
smoothed densities of the weights of evidence in cases and in controls.
}
\description{
Plot the cumulative frequency distributions in cases and in controls
}
\examples{
data(cleveland)
densities <- with(cleveland, Wdensities(y, posterior.p, prior.p))
plotcumfreqs(densities)

}
