% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotW}
\alias{plotW}
\title{plot log case/control density ratio against weight of evidence as a check that
the densities are mathematically consistent}
\usage{
plotW(densities, W)
}
\arguments{
\item{densities}{Adjusted densities computed by
\code{\link{Wdensities.fromraw}}.}

\item{W}{Weight of evidence. (natural logs)}
}
\value{
ggpplot of natural log case/control density ratio against weight of evidence
(should be a straight line of gradient 1 passing through the origin)
}
\description{
plot log case/control density ratio against weight of evidence as a check that
the densities are mathematically consistent
}
\examples{
data("cleveland") # load example dataset
W <- with(cleveland, weightsofevidence(posterior.p, prior.p))
densities.unadj <- Wdensities.unadjusted(cleveland$y, W)
densities.adj <- Wdensities.fromraw(densities.unadj)
plotW(densities.adj, W)

}
