% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdists.R
\name{recalibrate.p}
\alias{recalibrate.p}
\title{Recalibrate posterior probabilities}
\usage{
recalibrate.p(y, posterior.p)
}
\arguments{
\item{y}{Binary outcome label (0 for controls, 1 for cases).}

\item{posterior.p}{Vector of posterior probabilities.}
}
\value{
Recalibrated posterior probabilities.
}
\description{
Transforms posterior probabilities to logits, fits a logistic regression model
and returns the predictive probabilities from this model.
}
