% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdists.R
\name{tobits}
\alias{tobits}
\title{Convert from natural log units to bits}
\usage{
tobits(x)
}
\arguments{
\item{x}{Value expressed in natural log units (nats).}
}
\value{
Value expressed in bits.
}
\description{
Convert from natural log units to bits
}
\keyword{internal}
